swrenderer: First pass at an async rendering API
Introduce a new TargetPixelBuffer trait with async rendering API. To
start with, we focus on being able to fill a rectangle.

Introduce a RenderToAsyncBuffer and have it implement ProcessScene.
This allows us to work on the async API without affecting the existing
API.

Introduce a new SoftwareRenderer method:

``` pub async fn render_buffer(&self, buffer: &mut impl
TargetPixelBuffer) -> PhysicalRegion ```

This takes a TargetPixelBuffer and renders into it using the
RenderToAsyncBuffer scene processor.

Finally we modify WinitSoftwareRenderer so that the render method
checks for the `SLINT_ASYNC_RENDERING` environment variable, and if
found, uses the new API.