Use the `zephyr,touch` device tree entry for input handling
Zephyr have recently added this device tree entry to provide a common
interface for input handlers.

This means we can remove the overlay required for the i.MX RT1170
shield, and remove our modifications some of our modifications for the
native sim. It also means that we don't have to make device tree
modifications to support other boards in the future.